chore: updated code structure and better indentation in some files
