+++ backend/.env.example
@@ -1,5 +1,7 @@
-MONGO_URL=
+MONGOURL=
 CLIENT_ID=
 CLIENT_SECRET=
 SESSION_SECRET=
 FRONTEND_URL=
+EMAIL=
+PASSWORD=
