+++ backend/server/index.js
@@ -5,11 +5,8 @@ const dotenv = require('dotenv');
 const session = require("express-session");
 const passport = require("passport");
 const OAuth2Strategy = require("passport-google-oauth2").Strategy;
-const userdb = require("../model/userSchema");
+const userdb = require("./model/userSchema");
 const routes = require("./routes/index");
-
-// Importing Routers
-const mailRouter = require('./routes/mail');
 
 // Loading the env variables
 dotenv.config();
@@ -19,9 +16,6 @@ const PORT = 5000;
 const frontendUrl = process.env.FRONTEND_URL;
 const clientId = process.env.CLIENT_ID;
 const clientSecret = process.env.CLIENT_SECRET;
-
-// Loading routers on the app
-app.use('/api/mail', mailRouter)
 
 // Middleware
 app.use(
@@ -41,6 +35,9 @@ app.use(
     saveUninitialized: true,
   })
 );
+
+//Mounting of Router
+app.use("/api", routes);
 
 app.use(passport.initialize());
 app.use(passport.session());
@@ -87,6 +84,12 @@ passport.deserializeUser(async (id, done) => {
   }
 });
 
+// Basic route
+app.get("/", (req, res) => {
+  res.send("Hello from Express backend");
+});
+
+// Authentication Routes
 app.get(
   "/auth/google",
   passport.authenticate("google", { scope: ["profile", "email"] })
@@ -99,11 +102,6 @@ app.get(
     failureRedirect: `${frontendUrl}/login`,
   })
 );
-
-// Basic route
-app.get("/", (req, res) => {
-  res.send("Hello from Express backend");
-});
 
 app.get("/login/success", async (req, res) => {
   if (req.user) {
@@ -121,10 +119,6 @@ app.get("/logout", (req, res, next) => {
     res.redirect(`${frontendUrl}`);
   });
 });
-
-app.use("/", routes);
-
-const MONGO_URL = process.env.MONGO_URL;
 
 // MongoDB connection (add your MongoDB URL)
 mongoose.connect(process.env.MONGOURL)
