+++ backend/server/model/bookingSchema.js
@@ -20,4 +20,5 @@ const bookingSchema = new mongoose.Schema({
 }, { timestamps: true });
 
 const Booking = new mongoose.model('Booking', bookingSchema);
+
 module.exports = Booking;
