+++ backend/server/routes/venue.routes.js
@@ -1,17 +1,25 @@
 const express = require("express");
 const router = express.Router();
 const {
-  getVenueById,
+  // getVenueById, // TODO Complete this
   getAllVenues,
   createVenue,
   updateVenue,
   deleteVenue,
-} = require("../../controllers/venue.controllers");
+} = require("../controllers/venue.controllers");
 
+// Route to get all venues
 router.get("/admin/venues", getAllVenues);
+
+//Route to create venues
 router.post("/admin/venues", createVenue);
-router.get("/admin/venues/:venueId", getVenueById);
-router.post("/admin/venues/:venueId", updateVenue);
-router.post("/admin/venues/delete/:venueId", deleteVenue);
+
+// router.get("/admin/venues/:venueId", getVenueById); //TODO WRITE THIS CONTROLLER THIS IS CURRENTLY UNDEFINED
+
+// Route to update venues
+router.put("/admin/venues/:venueId", updateVenue);
+
+// Route to delete venues
+router.delete("/admin/venues/delete/:venueId", deleteVenue);
 
 module.exports = router;
